fix: update dao sql

+++ Database/Daos/UserFavoriteDao.cs
@@ -7,22 +7,22 @@ namespace Butterfly.Database.Daos
     {
         internal static void Insert(IQueryAdapter dbClient, int userId, int roomId)
         {
-            dbClient.RunQuery("INSERT INTO `user_favorite` (user_id, room_id) VALUES ('" + userId + "','" + roomId + "')");
+            dbClient.RunQuery("INSERT INTO `user_favorite` (`user_id`, `room_id`) VALUES ('" + userId + "','" + roomId + "')");
         }
 
         internal static void Delete(IQueryAdapter dbClient, int userId, int roomId)
         {
-            dbClient.RunQuery("DELETE FROM `user_favorite` WHERE user_id = '" + userId + "' AND room_id = '" + roomId + "'");
+            dbClient.RunQuery("DELETE FROM `user_favorite` WHERE `user_id` = '" + userId + "' AND `room_id` = '" + roomId + "'");
         }
 
         internal static void Delete(IQueryAdapter dbClient, int roomId)
         {
-            dbClient.RunQuery("DELETE FROM `user_favorite` WHERE room_id = '" + roomId + "'");
+            dbClient.RunQuery("DELETE FROM `user_favorite` WHERE `room_id` = '" + roomId + "'");
         }
 
         internal static DataTable GetAll(IQueryAdapter dbClient, int userId)
         {
-            dbClient.SetQuery("SELECT room_id FROM `user_favorite` WHERE user_id = '" + userId + "'");
+            dbClient.SetQuery("SELECT `room_id` FROM `user_favorite` WHERE `user_id` = '" + userId + "'");
             return dbClient.GetTable();
         }
     }
